[P4ADEV-4101] validateTaxonomyCategory API
